[*] rolled consensus back to the state before GRPC [*] reorganized stuff in Phorkus.Proto

+++ src/Phorkus.Core/Consensus/ConsensusContext.cs
@@ -20,7 +20,7 @@ namespace Phorkus.Core.Consensus
 
         public byte ViewNumber;
         public DateTime LastBlockRecieved;
-        public long MyIndex;
+        public int MyIndex;
         public readonly KeyPair KeyPair;
         public uint SignaturesAcquired;
